refactor(CountryTile): add 'highlight' prop to component
Add a new prop called 'highlight' to the component. When it's true, it will
change the tile's background and the "capital" text color.
