+++ src/components/CountryTile/styled-components.ts
@@ -4,8 +4,13 @@ import styled from 'styled-components/native';
 import Colors from 'styles/colors';
 import Fonts from 'styles/fonts';
 
-export const StyledTouchableOpacity = styled.TouchableOpacity`
-  background-color: ${() => Colors.platinum};
+import type { CountryTileProps } from './types';
+
+export const StyledTouchableOpacity = styled.TouchableOpacity<
+  Pick<CountryTileProps, 'highlight'>
+>`
+  background-color: ${(props) =>
+    props.highlight ? Colors.dartmouthGreen : Colors.platinum};
   border-radius: 10px;
   flex-direction: row;
   padding-top: 16px;
@@ -28,8 +33,10 @@ export const StyledCapitalContainerView = styled.View`
   padding-top: 2px;
 `;
 
-export const StyledCapitalText = styled.Text`
-  color: ${() => Colors.graniteGray};
+export const StyledCapitalText = styled.Text<
+  Pick<CountryTileProps, 'highlight'>
+>`
+  color: ${(props) => (props.highlight ? Colors.platinum : Colors.graniteGray)};
   font-family: ${() => Fonts.NunitoSansBold};
   font-size: 14px;
 `;
